(wip) split library into components
